Only require argparse when needed
`argparse` is available in the stdlib in all non-ancient Python versions (2.7+), and installing it is pointless because the stdlib copy always takes precedence. The name conflict, however, causes problems in some edge cases and makes installing packages depending on TransitLeastSquares impossible.

This patch adds an [environment marker](https://www.python.org/dev/peps/pep-0508/#environment-markers), so the library is only pulled in when `argparse` is not present.